Abort container build early.

+++ start_bonsai.py
@@ -36,16 +36,21 @@ def create_containers(container_name, path_key):
         raise exc
     # Without clean, bazel somehow won't update
     subprocess.call(["bazel", "clean"], cwd="./src")
-    subprocess.call(["bazel", "build", "frontend:server"], cwd="./src")
-    subprocess.call(["bazel", "build", "client:static"], cwd="./src")
+    if subprocess.call(["bazel", "build", "frontend:server"], cwd="./src") != 0:
+        raise RuntimeError("Frontend build failed")
+    if subprocess.call(["bazel", "build", "client:static"], cwd="./src") != 0:
+        raise RuntimeError("Client build failed")
     shutil.copyfile("src/bazel-out/local_linux-fastbuild/genfiles/frontend/server.bin", "docker/frontend-server.bin")
     shutil.copyfile(path_key, "docker/key.json")
     shutil.copyfile("/etc/ssl/certs/ca-bundle.crt", "docker/ca-bundle.crt")
     shutil.rmtree("docker/static", ignore_errors=True)
     os.mkdir("docker/static")
-    subprocess.call(["tar", "-xf", "src/bazel-out/local_linux-fastbuild/bin/client/static.tar", "-C", "docker/static"])
-    subprocess.call(["docker", "build", "-t", container_name, "-f", "docker/frontend", "./docker"])
-    os.remove('docker/key.json')
+    try:
+        subprocess.call(["tar", "-xf", "src/bazel-out/local_linux-fastbuild/bin/client/static.tar", "-C", "docker/static"])
+        if subprocess.call(["docker", "build", "-t", container_name, "-f", "docker/frontend", "./docker"]) != 0:
+            raise RuntimeError("Container build failed")
+    finally:
+        os.remove('docker/key.json')
 
 def deploy_containers_local(container_name):
     print("Running containers locally")
